requirements-azure.txt
openai>=1.51.0,<3
python-dotenv>=1.0.1,<2
gradio~=4.44.0
